refactor: 컴포넌트 구조 통합 및 폴더구조 컨벤션 적용
