+++ src/shared/component/Table/Body.tsx
@@ -1,7 +1,9 @@
-import type { TableColumns } from "@/shared/type/Table";
-import TableBody from "./TableElements/TableBody";
+import type { TableColumns } from "@/shared/type/table";
 import TableCell from "./TableElements/TableCell";
-import TableRow from "./TableElements/TableRow";
+import {
+  tableBodyStyle,
+  tableRowStyle,
+} from "./TableElements/TableElements.css";
 
 type BodyProps<T> = {
   rows: T[];
@@ -10,18 +12,18 @@ type BodyProps<T> = {
 
 const Body = <T,>({ rows, cols }: BodyProps<T>) => {
   return (
-    <TableBody>
+    <tbody className={tableBodyStyle}>
       {rows.map((row, idx) => (
         // api 연결 후, raw data에 고유 id값 추가 등의 방안으로 key 교체하기
-        <TableRow key={idx}>
+        <tr className={tableRowStyle} key={idx}>
           {cols.map((col) => (
-            <TableCell key={col.key as string} textAlign={col.justify}>
+            <TableCell key={col.key?.toString()} textAlign={col.justify}>
               {col.Cell(row)}
             </TableCell>
           ))}
-        </TableRow>
+        </tr>
       ))}
-    </TableBody>
+    </tbody>
   );
 };
