--- src/shared/component/Table/TableElements/TableBody.tsx
@@ -1,9 +1,0 @@
-import clsx from "clsx";
-import type { ComponentProps } from "react";
-import { tableBodyStyle } from "./TableElements.css";
-
-const TableBody = ({ className, ...props }: ComponentProps<"tbody">) => (
-  <tbody className={clsx(tableBodyStyle, className)} {...props} />
-);
-
-export default TableBody;
