--- src/shared/component/Table/TableElements/TableCaption.tsx
@@ -1,9 +1,0 @@
-import clsx from "clsx";
-import type { ComponentProps } from "react";
-import { tableCaptionStyle } from "./TableElements.css";
-
-const TableCaption = ({ className, ...props }: ComponentProps<"caption">) => (
-  <caption className={clsx(tableCaptionStyle, className)} {...props} />
-);
-
-export default TableCaption;
