--- src/shared/component/Table/TableElements/TableRow.tsx
@@ -1,9 +1,0 @@
-import clsx from "clsx";
-import type { ComponentProps } from "react";
-import { tableRowStyle } from "./TableElements.css";
-
-const TableRow = ({ className, ...props }: ComponentProps<"tr">) => (
-  <tr className={clsx(tableRowStyle, className)} {...props} />
-);
-
-export default TableRow;
